Getting Started
I created a html file which I use for displaying my solutions. The functions.js file contains all my written functions, which I will be use to create the demos for my solutions in a later stage. The style.css is very basic and contains the styling for my codeblocks.